requirements.txt
apturl==0.5.2
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
contourpy==1.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.7
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.10.0
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.40
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.4.0
kazam==1.4.5
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
more-itertools==8.10.0
mpmath==1.3.0
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.1.78
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.11.0
ply==3.11
prompt-toolkit==3.0.39
protobuf==3.12.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
reportlab==3.6.8
requests==2.25.1
RPi.GPIO==0.7.1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.12
systemd-python==234
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ultralytics==8.0.201
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.8
xdg==5
xkit==0.0.0
zipp==1.0.0
